change target of the GameLocation.performAction prefix
Stardew Valley has two overloaded implementations of
GameLocation.performAction. The first takes a string action and splits
it into sub-strings. It then calls another overload of the performAction
which takes a string[] action instead of a single string value. The
patch that we used splits the action string and is run at the higher
layer function.

Instead, change the target to patch the lower function that takes a
string[] list instead. This avoids having to duplicate the logic for
splitting action strings, and is more reliable in the case where Stardew
Valley directly calls the lower function.

It is unclear if this was the case for Stardew Valley prior to 1.6

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>