add try/catch protection to the InventoryMenu draw functions
While debugging the update for Stardew Valley 1.6, I found that the
InventoryMenu.draw patches were crashing. According to the modding wiki,
it is best practice for patches to catch all exceptions and default to
regular behavior if possible. For now, add try/catch blocks to these two
patches. I'll investigate adding similar exception handling to the other
patches in a follow-up change.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>