fix the Farmer_shiftToolbar_Patch for Stardew Valley 1.6
The Farmer_shiftToolbar_Patch Prefix function tries to check the
Game1.pickingTool variable and __instance.areAllItemsNull(). These
fields appear to have been removed in Stardew Valley 1.6

For Game1.pickingTool, there doesn't appear to be any equivalent
option, and the actual function we're patching no longer checks anything
like this.

For __instance.areAllItemsNull(), there is __instance.Items.HasAny(),
and the main function checks for this. Thus, align the check to match
whats in Stardew Valley 1.6

This appears to work in practice with the limited testing I've performed
so far.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>