stop using harmony.PatchAll() as recommended by modding wiki
The stardew valley modding wiki suggests that direct code patching via
harmony.Patch is preferred over using harmony annotations. The
annotations are possibly a bit less code, but they can be tricker to
follow. It also makes it more difficult to debug an issue with patching
during load of the mod.

Convert all the prefix and postfix functions to be part of an
ObjectPatches class, renaming them appropriately. Remove the annotations
and replace them with harmony.Patch invocations in the Entry function
for the mod.

This aligns with the best practice according to the modding wiki, even
if it is in a sense more verbose.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>